feat: make pseudocode remark plugin
