+++ gatsby-theme-oi-wiki/gatsby-config.js
@@ -87,6 +87,7 @@ module.exports = {
         remarkPlugins: [
           require('remark-math'),
           esmRequire('remark-details').default,
+          require('remark-pseudocodejs'),
           [require('@mgtd/remark-shiki').remarkShiki, {
             semantic: false,
             theme: 'css-variables',
@@ -94,9 +95,8 @@ module.exports = {
           }],
         ],
         rehypePlugins: [
-          require('./plugins/rehype-pseudocodejs'),
           mathRehype,
-          require('./plugins/rehype-codeblock'),
+          require('rehype-codeblock'),
         ],
         // extensions: ['.mdx', '.md'],
       },
