+++ gatsby-theme-oi-wiki/plugins/remark-pseudocodejs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "rehype-pseudocodejs",
+  "name": "remark-pseudocodejs",
   "private": true,
   "version": "1.0.0",
   "description": "",
@@ -7,7 +7,6 @@
   "author": "Xiaodai Dai",
   "license": "MIT",
   "dependencies": {
-    "hast-util-to-string": "^1.0.4",
     "pseudocode": "^2.2.0",
     "unist-util-visit": "^2.0.3"
   }
